Python/requirements.txt
pymysql
pyautogui
pyperclip
pynput
python-dotenv
opencv-python
pillow
numpy